% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors_carp.R
\name{get_cluster_labels}
\alias{get_cluster_labels}
\alias{get_cluster_labels.CARP}
\alias{get_cluster_centroids}
\alias{get_cluster_centroids.CARP}
\alias{get_clustered_data}
\alias{get_clustered_data.CARP}
\title{Get Clustering Results for \code{CARP}}
\usage{
get_cluster_labels(x, ...)

\method{get_cluster_labels}{CARP}(x, ..., percent, k)

get_cluster_centroids(x, ...)

\method{get_cluster_centroids}{CARP}(x, ..., percent, k, refit = TRUE)

get_clustered_data(x, ...)

\method{get_clustered_data}{CARP}(x, ..., percent, k, refit = TRUE)
}
\arguments{
\item{x}{An object of class \code{CARP} as produced by \code{\link{CARP}}}

\item{...}{Additional arguments - if any are provided, an error is signalled.}

\item{percent}{A number between 0 and 1, giving the regularization level (as
a fraction of the final regularization level used) at which to
get cluster labels.}

\item{k}{The desired number of clusters. If no iteration with exactly this
many clusters is found, the first iterate with fewer than \code{k}
clusters is used.}

\item{refit}{Should "naive" centroids (\code{TRUE}) or the actual centroids
estimated by convex clustering be used? The default (\code{refit = TRUE})
centroids returned are actual centroids (mean) of all elements
assigned to that cluster; if \code{refit = FALSE}, the \eqn{\hat{U}}
from the convex clustering problem is used. Due to the global
shrinkage imposed, these clusters are more "shrunk together" than
the naive clusters.}
}
\description{
\code{get_cluster_labels} returns a factor vector of cluster labels.
\code{get_clustered_data} returns a matrix (with the same dimensions and names
as the original data), but with the values for each observation replaced by
its "estimated" value (\emph{i.e.}, the appropriate cluster centroid).
\code{get_cluster_centroids} returns a \code{k}-by-\code{p} matrix of cluster
centroids, with the same column names as the original data.
}
\details{
\code{get_clustered_data} and \code{get_cluster_centroids} return
centroids on the original scale of the data, independent of any pre-processing
flags passed to \code{CARP}. Note that exactly one of \code{percent} and
\code{k} must be supplied to each function.
}
\examples{
carp_fit <- CARP(presidential_speech)

# Get clustering results from 50\% along the path
get_cluster_labels(carp_fit, percent = 0.5)

# Get labels corresponding to the 3 cluster solution
get_cluster_labels(carp_fit, k = 3)

# Get 3 cluster centroids
get_cluster_centroids(carp_fit, k = 3)

# Get the clustered estimates for k = 3 clusters
get_clustered_data(carp_fit, k = 3)
}
