% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors_cbass.R
\name{get_cluster_labels.CBASS}
\alias{get_cluster_labels.CBASS}
\alias{get_cluster_centroids.CBASS}
\alias{get_clustered_data.CBASS}
\title{Get Clustering Results for \code{CBASS}}
\usage{
\method{get_cluster_labels}{CBASS}(x, ..., percent, k.row, k.col, type = c("row", "col"))

\method{get_cluster_centroids}{CBASS}(x, ..., percent, k.row, k.col, refit = TRUE)

\method{get_clustered_data}{CBASS}(x, ..., percent, k.row, k.col, refit = TRUE)
}
\arguments{
\item{x}{An object of class \code{CARP} as produced by \code{\link{CBASS}}}

\item{...}{Additional arguments - if any are provided, an error is signalled.}

\item{percent}{A number between 0 and 1, giving the regularization level (as
a fraction of the final regularization level used) at which to
get cluster labels.}

\item{k.row}{The desired number of row clusters}

\item{k.col}{The desired number of column clusters}

\item{type}{For \code{get_cluster_labels}, which set of labels to return -
row (observation) or column (feature)}

\item{refit}{Should "naive" centroids (\code{TRUE}) or the actual centroids
estimated by convex clustering be used? The default (\code{refit = TRUE})
centroids returned are actual centroids (mean) of all elements
assigned to that cluster; if \code{refit = FALSE}, the \eqn{\hat{U}}
from the convex biclustering problem is used. Due to the global
shrinkage imposed, these clusters are more "shrunk together" than
the naive clusters.}
}
\description{
\code{get_cluster_labels} returns a factor vector of cluster labels.
\code{get_cluster_centroids} returns a \code{k1}-by-\code{k2} matrix with the
estimated centroid of the \code{k1}-th row cluster and the \code{k2}-th
column cluster.
\code{get_clustered_data} returns a matrix (with the same dimensions and names
as the original data), but with the values for each row replaced by
its "estimated" value (\emph{i.e.}, the appropriate cluster centroid).
}
\details{
\code{get_clustered_data} returns centroids on the original scale of
the data, independent of any pre-processing flags passed to \code{CBASS}.
Note that exactly one of \code{percent}, \code{k.row}, \code{k.col}
must be supplied and that that \code{k.row} (if suppplied) will be
used even if \code{type = "col"} and \emph{vice versa}.
}
\examples{
cbass_fit <- CBASS(presidential_speech)

# Get row clustering results from 50\% along the path
get_cluster_labels(cbass_fit, percent = 0.5)

# Get column clustering corresponding to the 3 cluster solution
get_cluster_labels(cbass_fit, k.col = 3, type = "col")

# Get row clustering corresponding to the 3 column clusters
get_cluster_labels(cbass_fit, k.col = 3, type = "row")

# Get cluster centroids partially down the path
get_cluster_centroids(cbass_fit, percent = 0.5)

# Get clustered data
image(get_clustered_data(cbass_fit, k.row = 2))
}
