% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{authors}
\alias{authors}
\title{Word Count Data from Four English-Language Authors}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 841 rows and 69 columns.
}
\usage{
authors
}
\description{
This data set (\eqn{n=841, p = 69}) consists of counts of common words
appearing in texts written by four popular English-language authors 
(Jane Austen, Jack London, William Shakespeare, and John Milton).
The row names are the authors (true cluster labels) and the column
names are the words (slightly processed).
}
\keyword{datasets}
