% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{carp.control}
\alias{carp.control}
\title{Control for CARP fits}
\usage{
carp.control(obs.labels = NULL, var.labels = NULL, X.center = TRUE,
  X.scale = FALSE, phi = 0.001, rho = 1, weights = NULL, k = NULL,
  weight.dist = "euclidean", weight.dist.p = 2, ncores = as.integer(1),
  max.iter = as.integer(1e+06), burn.in = as.integer(50),
  alg.type = "carpviz", t = 1.05, npcs = as.integer(4), ...)
}
\arguments{
\item{obs.labels}{a vector of length n containing observations (row) labels}

\item{var.labels}{a vector of length p containing variable (column) labels}

\item{X.center}{A logical. Should X be centered?}

\item{X.scale}{A logical. Should X be scaled?}

\item{phi}{A positive numner used for scaling in RBF kernel}

\item{rho}{A positive number for augmented lagrangian. Not advisable to change.}

\item{weights}{A vector of positive number of length choose(n,2).}

\item{k}{an integer >= 1. The number of neighbors used to create sparse weights}

\item{weight.dist}{a string indicating the distance metric used to calculate weights.}

\item{weight.dist.p}{The power of the Minkowski distance, if used.}

\item{ncores}{an integer >= 1. The number of cores to use.}

\item{max.iter}{an integer. The maximum number of CARP iterations.}

\item{burn.in}{an integer. The number of initial iterations at a fixed
value of (small) lambda_k}

\item{alg.type}{Which CARP algorithm to perform. Choices are 'carpviz'
and 'carp;}

\item{t}{a number greater than 1. The size of the multiplicitive
regularization parameter update. Typical values are: 1.1, 1.05, 1.01, 1.005
Determines observation pairs' fusion amount.}

\item{npcs}{A integer >= 2. The number of principal components to compute
for path visualization.}

\item{...}{unused generic arguements.}
}
\value{
a list containing the CARP options.
}
\description{
Parameters for various CARP fitting options
}
