% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cbass.control}
\alias{cbass.control}
\title{Control for CBASS fits}
\usage{
cbass.control(obs.labels = NULL, var.labels = NULL,
  X.center.global = TRUE, rho = 1, phi = 0.1, weights.obs = NULL,
  weights.var = NULL, obs.weight.dist = "euclidean",
  obs.weight.dist.p = 2, var.weight.dist = "euclidean",
  var.weight.dist.p = 2, k.obs = NULL, k.var = NULL, t = 1.01,
  ncores = as.integer(1), max.iter = as.integer(1e+06),
  burn.in = as.integer(50), alg.type = "cbassviz", npcs = as.integer(4))
}
\arguments{
\item{obs.labels}{a vector of length n.obs containing observations (row) labels}

\item{var.labels}{a vector of length p.var containing variable (column) labels}

\item{X.center.global}{a logical. If TRUE, the global mean of X is removed.}

\item{rho}{A positive number for augmented lagrangian. Not advisable to change.}

\item{phi}{A positive numner used for scaling in RBF kernel}

\item{weights.obs}{A vector of positive number of length choose(n.obs,2).
Determines observation pair fusions weight.}

\item{weights.var}{A vector of positive number of length choose(p.var,2).
Determines variable pair fusions weight.}

\item{obs.weight.dist}{a string indicating the distance metric used to calculate
observation weights}

\item{obs.weight.dist.p}{The power of the Minkowski distance, if used for
observation weights.}

\item{var.weight.dist}{a string indicating the distance metric used to calculate
variable weights}

\item{var.weight.dist.p}{The power of the Minkowski distance, if used for
variable weights.}

\item{k.obs}{an integer >= 1. The number of neighbors used to create sparse
observation weights}

\item{k.var}{an integer >= 1. The number of neighbors used to create sparse
variable weights}

\item{t}{a number greater than 1. The size of the multiplicitive
regularization parameter update. Typical values are: 1.1, 1.05, 1.01, 1.005.
Not used CBASS-VIZ algorithms.}

\item{ncores}{an integer >= 1. The number of cores to use.}

\item{max.iter}{an integer. The maximum number of CARP iterations.}

\item{burn.in}{an integer. The number of initial iterations at a fixed
value of (small) lambda_k}

\item{alg.type}{Which CARP algorithm to perform. Choices are 'cbassviz'
and 'cbass';}

\item{npcs}{A integer >= 2. The number of principal components to compute
for path visualization.}

\item{...}{unused additional arguements}
}
\value{
a list of CBASS parameters
}
\description{
Parameters for various CBASS fitting options
}
