% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CBASS}
\alias{CBASS}
\title{Compute CBASS solution path}
\usage{
CBASS(X, t = NULL, verbose = 1, interactive = TRUE, static = TRUE,
  control = NULL, ...)
}
\arguments{
\item{X}{A n.obs by p.var matrix with rows the observations and columns the variables}

\item{verbose}{either 0,1,or 2. Higher numbers indicate more verbosity while
computing.}

\item{interactive}{A logical. Should an interactive heatmap be returned?}

\item{static}{A logical. Should observation and variable dendrograms
be returned?}

\item{control}{a list of CBASS control arguements; see \code{cbass.control}}

\item{...}{additional arguments passed to \code{cbass.control}}
}
\value{
X the original data matrix

cbass.cluster.path.obs the CBASS observation solution path

cbass.cluster.path.var the CBASS variable solution path

cbass.dend.obs if static==TRUE, a dendrogam representation of the
observation clustering solution path

cbass.dend.var if static==TRUE, a dendrogam representation of the
variable clustering solution path

phi.obs A positive numner used for scaling in observation RBF kernel

phi.var A positive numner used for scaling in variable RBF kernel

k.obs an integer >= 1. The number of neighbors used to create sparse
observation weights

k.var an integer >= 1. The number of neighbors used to create sparse
variable weights

burn.in an integer. The number of initial iterations at a fixed
value of (small) lambda_k

alg.type Which CARP algorithm to perform. Choices are 'cbassviz'
and 'cbass'

interactive A logical. Should an interactive heatmap be returned?

static A logical. Should observation and variable dendrograms
be returned?

obs.labels a vector of length n.obs containing observations (row) labels

var.labels a vector of length p.var containing variable (column) labels

X.center.global a logical. If TRUE, the global mean of X is removed.
}
\description{
\code{CBASS} returns a fast approximation to the Convex BiClustering
solution path along with visualizations such as dendrograms and
heatmaps. Visualizations may be static, interactive, or both.
}
\details{
\code{CBASS} solves the Convex Biclustering problem via
Algorithmic Regularization Paths. A seqeunce of biclustering
solutions is returned along with several visualizations.
}
\examples{
\dontrun{
library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech
CBASS(
X=Xdat
) -> cbass.fit
}
}
