% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Clustering.CBASS}
\alias{Clustering.CBASS}
\title{Get biclustering solution from a CBASS object}
\usage{
\method{Clustering}{CBASS}(x, k.obs = NULL, k.vars = NULL, percent = NULL,
  ...)
}
\arguments{
\item{x}{A CARP object returned by \code{CARP}}

\item{k.obs}{An interger between 1 and \code{n.obs}. The number of unique
observation clusters.}

\item{k.vars}{An interger between 1 and \code{p.vars}. The number of unique
variable clusters.}

\item{percent}{A number between 0 and 1. The percent of regularization at
which to cut the path.}

\item{...}{Unused additional generic arguements}
}
\value{
A list with elements
\describe{
\item{\code{clustering.assignment.obs}}{
A vector of observation cluster labels of length \code{n.obs}.
}
\item{\code{clustering.assignment.vars}}{
A vector of variable cluster labels of length \code{p.vars}.
}
\item{\code{cluster.mean.matrix}}{
The cluster mean matrix of size \code{n.obs} by \code{p.vars}.
}
}
}
\description{
Returns (observation and variable) cluster labels and
cluster mean matrix at a point along the CBASS path.
}
\details{
Passing either the desired number of clusters (\code{k}) or the percent
regularization (\code{percent}) returns the clustering assignment
and cluster mean matrix at the specific point along the CBASS path.
}
\examples{
library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech$X[1:10,1:4]
}
