% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Clustering}
\alias{Clustering}
\title{Method for returning CARP and CBASS clustering solutions}
\usage{
Clustering(x, ...)
}
\arguments{
\item{x}{a CARP or CBASS object}

\item{...}{additional arguements passed to \code{Clustering.CARP} or
\code{Clustering.CBASS}}
}
\value{
CARP clustering solutions or CBASS biclustering solutions
}
\description{
See \code{Clustering.CARP} and \code{Clustering.CBASS} for details
}
