% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustRviz-package.R
\docType{package}
\name{clustRviz}
\alias{clustRviz}
\title{clustRviz: Efficient Computation and Dynamic Visualizations for Convex Clustering and BiClustering}
\description{
Implementation of the CARP and CBASS algorithms for fast convex clustering.
See the package vignettes and accompanying paper for details of the algorithm
and comparisons to existing methods.
}
