% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{clustRviz_logger_level}
\alias{clustRviz_logger_level}
\alias{clustRviz_logging}
\title{\code{ClustRViz} Package Logging Functionality}
\usage{
clustRviz_logger_level(
  level = c("ERROR", "WARNING", "MESSAGE", "INFO", "DEBUG")
)
}
\arguments{
\item{level}{The desired new log level. Available levels are \itemize{
    \item \code{ERROR} - corresponding to \code{base::stop};
    \item \code{WARNING} - corresponding to \code{base::warning};
    \item \code{MESSAGE} - corresponding to \code{base::message};
    \item \code{INFO}; and
    \item \code{DEBUG.}
} If omitted, the log level is not changed (and the current level is still
returned invisibly.) See below for details about the different levels.}
}
\value{
The previous log level (invisibly).
}
\description{
Control the global logging level for the \code{clustRviz} package.
}
\details{
The \code{clustRviz} package has a multi-level logging system, with a single
    global log level; (which applies to both \code{R} and \code{C++} level
    functionality.) the levels are, in decreasing order, \code{ERROR},
    \code{WARNING}, \code{MESSAGE} (default), \code{INFO}, \code{DEBUG}.

    To change the amount of output from the \code{clustRviz} package, the
    \code{clustRviz_logger_level} function can be used to adjust the global
    log level. The \code{INFO} and \code{DEBUG} levels can be quite verbose
    and may significantly slow down the package.
}
\examples{
# Switch to INFO level and fit somewhat loudly
clustRviz_logger_level("INFO")
CARP(presidential_speech)

# Return to default behavior
clustRviz_logger_level("MESSAGE")
# Will give a few messages about phases of fitting
CARP(presidential_speech)

# If we switch to ERRORs only, no progress message
clustRviz_logger_level("ERROR")
CARP(presidential_speech)

# Return to default
clustRviz_logger_level("MESSAGE")

# Normally this prints messages as the paths are calculated
CARP(presidential_speech)
# But it can be suppressed using standard R functions
suppressMessages(CARP(presidential_speech))
}
