% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{convex_clustering}
\alias{convex_clustering}
\title{Compute Convex Clustering Solution Path on a User-Specified Grid}
\usage{
convex_clustering(
  X,
  ...,
  lambda_grid,
  weights = sparse_rbf_kernel_weights(k = "auto", phi = "auto", dist.method =
    "euclidean", p = 2),
  X.center = TRUE,
  X.scale = FALSE,
  norm = 2,
  impute_func = function(X) {     if (anyNA(X))          missForest(X)$ximp     else X
    },
  status = (interactive() && (clustRviz_logger_level() \%in\% c("MESSAGE", "WARNING",
    "ERROR")))
)
}
\arguments{
\item{X}{The data matrix (\eqn{X \in R^{n \times p}}{X}): rows correspond to
the observations (to be clustered) and columns to the variables (which
will not be clustered). If \code{X} has missing values - \code{NA} or
\code{NaN} values - they will be automatically imputed.}

\item{...}{Unused arguements. An error will be thrown if any unrecognized
arguments as given. All arguments other than \code{X} must be given
by name.}

\item{lambda_grid}{A user-supplied set of \eqn{\lambda}{lambda} values at which
to solve the convex clustering problem. These must be strictly
positive values and will be automatically sorted internally.}

\item{weights}{One of the following: \itemize{
\item A function which, when called with argument \code{X},
      returns an b-by-n matrix of fusion weights.
\item A matrix of size n-by-n containing fusion weights
}}

\item{X.center}{A logical: Should \code{X} be centered columnwise?}

\item{X.scale}{A logical: Should \code{X} be scaled columnwise?}

\item{norm}{Which norm to use in the fusion penalty? Currently only \code{1}
and \code{2} (default) are supported.}

\item{impute_func}{A function used to impute missing data in \code{X}. By default,
the \code{\link[missForest]{missForest}} function from the
package of the same name is used. This provides a flexible
potentially non-linear imputation function. This function
has to return a data matrix with no \code{NA} values.
Note that, consistent with base \code{R}, both \code{NaN}
and \code{NA} are treaded as "missing values" for imputation.}

\item{status}{Should a status message be printed to the console?}
}
\value{
An object of class \code{convex_clustering} containing the following elements (among others):
        \itemize{
        \item \code{X}: the original data matrix
        \item \code{n}: the number of observations (rows of \code{X})
        \item \code{p}: the number of variables (columns of \code{X})
        \item \code{X.center}: a logical indicating whether \code{X} was centered
                               column-wise before clustering
        \item \code{X.scale}: a logical indicating whether \code{X} was scaled
                              column-wise before centering
        \item \code{weight_type}: a record of the scheme used to create
                                  fusion weights
        \item \code{U}: a tensor (3-array) of clustering solutions
        }
}
\description{
\code{convex_clustering} calculates the convex clustering solution path
at a user-specified grid of lambda values (or just a single value). It is,
in general, difficult to know a useful set of lambda values \emph{a priori},
so this function is more useful for timing comparisons and methodological
research than applied work.
}
\details{
Compared to the \code{\link{CARP}} function, the returned object
is much more "bare-bones," containing only the estimated \eqn{U} matrices,
and no information used for dendrogram or path visualizations.
}
\examples{
clustering_fit <- convex_clustering(presidential_speech[1:10,1:4], lambda_grid = 1:100)
print(clustering_fit)
}
