% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DenseWeights}
\alias{DenseWeights}
\title{Compute dense distanced-based gaussian kernel weights for use in CARP or CBASS}
\usage{
DenseWeights(X, phi = 1, method = "euclidean")
}
\arguments{
\item{X}{an n by p matrix with rows the observations and columns the variables.}

\item{phi}{a number. Scaling parameter used in the gaussian kernel.}

\item{method}{a string. Passed to the \code{dist} function. See \code{?dist}}
}
\value{
dense.weights a numeric vector of weights
}
\description{
Compute dense distanced-based gaussian kernel weights for use in CARP or CBASS
}
