% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MinKNN}
\alias{MinKNN}
\title{Compute the minimum number of nearest neighbors required to fully
cluster all observations}
\usage{
MinKNN(X, dense.weights)
}
\arguments{
\item{X}{an n.obs x p.vars matrix}

\item{dense.weights}{a vector vector of dense weights such as returned by
\code{DenseWeights}}
}
\value{
k an integer. The smallest number of nearest neighbors required to
fully cluster all observations.
}
\description{
Compute the minimum number of nearest neighbors required to fully
cluster all observations
}
