% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{clustRviz_options}
\alias{clustRviz_options}
\alias{clustRviz_reset_options}
\title{\code{ClustRViz} Options}
\usage{
clustRviz_options(...)

clustRviz_reset_options()
}
\arguments{
\item{...}{Options (to be passed by name). See below for available options.}
}
\description{
Advanced control of algorithmic options for \code{\link{CARP}} and \code{\link{CBASS}}.
The \code{clustRviz_reset_options} function returns options to "factory-fresh"
settings.
}
\details{
The following options can be set by name:\itemize{
  \item \code{epsilon} The initial step size (fixed during the "burn-in" period)
  \item \code{stopping_threshold}: Stopping threshold to use for \code{convex_clustering},
                                   \code{convex_biclustering}, \code{CARP(exact = TRUE)},
                                   and \code{CBASS(exact = TRUE)}. By default,
                                   this is set to \code{1e-10} - a very conservative
                                   threshold: making it larger can significantly
                                   improve performance
  \item \code{max_iter} An integer: the maximum number of iterations to perform
  \item \code{max_inner_iter} An integer: the maximum number of iterations for
                              the iterative solvers to perform at a single
                              value of \eqn{\lambda}{\lambda}.
  \item \code{burn_in} An integer: the number of initial iterations at a fixed
                      (small) value of \eqn{\gamma}
  \item \code{viz_initial_step} The initial (large) step size used in back-tracking
                                (\code{CARP-VIZ} and \code{CBASS-VIZ}) algorithms.
  \item \code{viz_small_step} The secondary (small) step size used in back-tracking
                              (\code{CARP-VIZ} and \code{CBASS-VIZ}) algorithms.
  \item \code{viz_max_inner_iter} The maximum number of iterations to perform
                                  in the inner loop of back-tracking (\code{CARP-VIZ}
                                  and \code{CBASS-VIZ}) algorithms.
  \item \code{keep} \code{\link{CARP}} and \code{\link{CBASS}} keep every
                    \code{keep}-th iteration even if no fusions are detected.
                    Increasing this parameter may improve performance, at
                    the expense of returning a finer grid.
  \item \code{rho} For advanced users only (not advisable to change): the penalty
                   parameter used for the augmented Lagrangian.
  \item \code{keep_debug_info}: Should additional debug info (currently only the V-path)
                                be kept?
}
}
