% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_carp.R
\name{plot.CARP}
\alias{plot.CARP}
\title{Plot function for CARP objects}
\usage{
\method{plot}{CARP}(x, type = c("dendrogram", "path", "interactive"),
  axis = c("PC1", "PC2"), dend.branch.width = 2,
  dend.labels.cex = 0.6, percent = 1, max.nclust = 9,
  min.nclust = 1, ...)
}
\arguments{
\item{x}{a CARP object returned by \code{CARP}}

\item{type}{a string specifying the type of plot to produce. 'dendrogram'
produces the static cluster dendrogram; 'path' produces the static
cluster path; and 'interactive' produces an interactive visualization of
both the cluster path and dendrogram}

\item{axis}{a character vector of length two with elements as 'PC1','PC2',..
etc. Specifics which principal component axis to display for the 'path'
visualization}

\item{dend.branch.width}{a positive number. Line width on dendrograms.}

\item{dend.labels.cex}{a positive number. Label size on dendrograms.}

\item{percent}{a number between 0 and 1. Specifies how far along the
CARP path the 'path' visualization should display.}

\item{max.nclust}{a positive integer. The maximum number of clusters
to display in the interactive plot.}

\item{min.nclust}{a positive value. The minimum number of clusters to
display in the interactive plot.}

\item{...}{Unused additional generic arguements}
}
\description{
\code{plot.CARP} displays both static and interactive visualizations of
the CARP clustering solution path, including dendrograms and clustering
paths
}
\details{
Possible visualizations of the CARP solution path are: (i) a static
dendrogram representing the clustering solution path at various
levels of regularization; (ii) a static clustering path representing
the coalescence of observations projected via PCA; and (iii) a
interactive dendorgram and clustering path visualization, showing
how the clustering solutions change with increased reguarliazation.
The first interactive tab shows a movie of real-time cluster solutions
while the second tabs shows the cluster solution for various numbers
of clusters
}
\examples{
\dontrun{
library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech[1:10,1:4]
carp.fit <- CARP(X=Xdat)
plot(carp.fit,type='interactive')
}
}
