% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_carp.R
\name{plot.CARP}
\alias{plot.CARP}
\title{Visualize the results of Convex Clustering (\code{CARP})}
\usage{
\method{plot}{CARP}(
  x,
  ...,
  type = c("dendrogram", "path", "heatmap"),
  dynamic = FALSE,
  interactive = FALSE,
  axis = c("PC1", "PC2"),
  percent,
  k,
  percent.seq = seq(0, 1, 0.01),
  slider_y = -0.3,
  refit = FALSE
)
}
\arguments{
\item{x}{An object of class \code{CARP} as returned by \code{\link{CARP}}}

\item{...}{Additional arguments, which are handled differently for different
values of \code{type}.\itemize{
\item When \code{type} is \code{"path"}, the presence of
      unknown arguments triggers an error;
\item when \code{type == "dendrogram"}
      \code{...} is forwarded to \code{\link[stats]{dendrogram}}; and
\item when \code{type == "heatmap"} and \code{interactive == TRUE},
      \code{...} is forwarded to \code{\link[heatmaply]{heatmaply}}.
} See the documentation of \code{\link[ggplot2]{ggplot}} (for \code{interactive == FALSE})
or \code{\link[plotly]{plotly}} (for \code{interactive == TRUE}) for details about additional
supported arguments to the corresponding plot \code{type}.}

\item{type}{A string indicating the type of visualization to show (see details above).}

\item{dynamic}{A logical scalar.Should the resulting animation be dynamic (animated) or not?
If \code{TRUE}, a dynamic visualization which varies along the CARP solution path at a
grid given by \code{percent.seq} is produced. If \code{FALSE}, a fixed visualization at a single
solution (determined by either \code{percent} or \code{k} if supplied) is produced.}

\item{interactive}{A logical scalar. Should the resulting animation be interactive or not?
If \code{TRUE}, an interactive visualization is produced by javascript(\code{\link{plotly}}).
If \code{FALSE}, a non-interactive visualization is produced by \code{\link[ggplot2]{ggplot}}.}

\item{axis}{A character vector of length two indicating which features or principal
components to use as the axes in the \code{type = "path"} visualization.
Currently only features like \code{"PC1"} or \code{"PC2"} (indicating
the first principal component projections) are supported.}

\item{percent}{A number between 0 and 1, giving the regularization level (as
a fraction of the final regularization level used) at which to
assign clusters in the static (\code{type = "dendrogram"} or \code{type = "path"})
plots.}

\item{k}{An integer indicating the desired number of clusters to be displayed
in the static plots. If no \code{CARP} iteration with exactly this
many clusters is found, the first iterate with fewer than \code{k}
clusters is used.}

\item{percent.seq}{A grid of values of percent along which to generate dynamic
visualizations (if dynamic == TRUE)}

\item{slider_y}{A number to adjust the slider's vertical position for
interactive dendrogram and interactive heatmap plots
(ignored for other plot types).}

\item{refit}{A logical scalar. Should "naive" centroids (TRUE) or the
actual centroids estimated by convex clustering be used?
When the default refit = FALSE, the estimated U from the convex
clustering problem is used. The refit = TRUE returns actual
centroids (mean) of all elements assigned to that cluster;
Due to the global shrinkage imposed, these clusters are
more "shrunk together" than the naive clusters. Only for the
heatmap plots. (ignored for other plot types).}
}
\value{
The value of the return type depends on the \code{interactive} and \code{dynamic} arguments:\itemize{
  \item if \code{interactive = FALSE} and \code{dynamic = FALSE}, an object of class \code{\link[ggplot2]{ggplot}}
        is returned;
  \item if \code{interactive = FALSE} and \code{dynamic = TRUE}, an object of class \code{\link[gganimate:gganimate-package]{gganim}}
        is returned;
  \item if \code{interactive = TRUE}, an object of class \code{\link[plotly]{plotly}}
        is returned.
} All the plots can be plotted directly (by invoking its print method) or further
  manipulated by the user.
}
\description{
\code{plot.CARP} provides a range of ways to visualize the results of convex
clustering, including: \itemize{
\item A dendrogram, illustrating the nested cluster hierarchy inferred from
      the convex clustering solution path (\code{type = "dendrogram"});
\item A static path plot, showing the coalescence of the estimated cluster centroids
      at a fixed value of the regularization parameter is increased (\code{type = "path"});
\item A \code{\link[gganimate]{gganimate}} plot, showing the coalescence of the
      estimated cluster centroids as the regularization parameter is increased
      (\code{dynamic = TRUE})
}
}
\examples{
carp_fit <- CARP(presidential_speech)
plot(carp_fit, type = "path")
plot(carp_fit, type = "dendrogram")
plot(carp_fit, type = "heatmap")
plot(carp_fit, type = "heatmap", dynamic = TRUE)
\dontrun{
plot(carp_fit, type='heatmap', interactive=TRUE)
}
}
