% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PlotWeightGraph}
\alias{PlotWeightGraph}
\title{Plot graph induced by weights}
\usage{
PlotWeightGraph(weights, nobs, edge.labels = TRUE, obs.labels = NULL, ...)
}
\arguments{
\item{weights}{a vector of weights such as returned by \code{SparseWeights}}

\item{nobs}{the number of observations being clustered}

\item{edge.labels}{a logical. Should weight values be displayed along edges?}

\item{obs.labels}{a vector of observation labels}

\item{...}{additional parameters passed to qgraph::qgraph}
}
\description{
Plot graph induced by weights
}
