% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{presidential_speech}
\alias{presidential_speech}
\title{Log transformed word count of presidential speeches}
\format{A list with elements
\describe{
  \item{labels}{a vector of row labels, one for each presidents}
  \item{X}{a 44 by 75 data matrix of log transformed word counts with
  the stemmed words the column names}
  ...
}}
\source{
\url{http://www.presidency.ucsb.edu}
}
\usage{
presidential_speech
}
\description{
A dataset of the top 75 most variable log-transformed word counts for
each US president aggregated over several speeches
(Inaugural, State of the Union, etc.).
Stop words have been removed and words have been stemmed.
}
\keyword{datasets}
