% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.CARP}
\alias{print.CARP}
\title{Print CARP Summary}
\usage{
\method{print}{CARP}(x, ...)
}
\arguments{
\item{x}{a CARP object returned by \code{CARP}}

\item{...}{Unused additional generic arguements}
}
\description{
Prints a brief descripton of a fitted carp object.
}
\details{
Reports number of observations and variables of dataset, any preprocessing
done by the \code{CARP} function, regularization weight information,
the type of CARP algorithm performed, and the visualizations returned.
}
\examples{
library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech[1:10,1:4]
carp.fit <- CARP(X=Xdat)
print(carp.fit)
}
