% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.CBASS}
\alias{print.CBASS}
\title{Print CBASS Summary}
\usage{
\method{print}{CBASS}(x, ...)
}
\arguments{
\item{x}{a CBASS object returned by \code{CARP}}

\item{...}{Unused additional generic arguements}
}
\description{
Prints a brief descripton of a fitted cbass object.
}
\details{
Reports number of observations and variables of dataset, any preprocessing
done by the \code{CBASS} function, regularization weight information,
the type of CBASS algorithm performed, and the visualizations returned.
}
\examples{
library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech[1:10,1:4]
cbass.fit <- CBASS(X=Xdat)
print(cbass.fit)
}
