% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{print.ConvexBiClustering}
\alias{print.ConvexBiClustering}
\title{Print \code{\link{convex_biclustering}} Results}
\usage{
\method{print}{ConvexBiClustering}(x, ...)
}
\arguments{
\item{x}{an object of class \code{convex_biclustering} as returned by
\code{\link{convex_biclustering}}}

\item{...}{Additional unused arguments}
}
\description{
Prints a brief descripton of a fitted \code{convex_biclustering} object.
}
\details{
Reports number of observations and variables of dataset, any preprocessing
done by the \code{\link{convex_biclustering}} function, and regularization
details.
}
\examples{
bicluster_fit <- convex_biclustering(presidential_speech, lambda_grid = 1:5)
print(bicluster_fit)
}
