% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{print.ConvexClustering}
\alias{print.ConvexClustering}
\title{Print \code{\link{convex_clustering}} Results}
\usage{
\method{print}{ConvexClustering}(x, ...)
}
\arguments{
\item{x}{an object of class \code{convex_clustering} as returned by
\code{\link{convex_clustering}}}

\item{...}{Additional unused arguments}
}
\description{
Prints a brief descripton of a fitted \code{convex_clustering} object.
}
\details{
Reports number of observations and variables of dataset, any preprocessing
done by the \code{\link{convex_clustering}} function, and regularization
details.
}
\examples{
cluster_fit <- convex_clustering(presidential_speech, lambda_grid = 1:5)
print(cluster_fit)
}
