% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carp.R
\name{print.CARP}
\alias{print.CARP}
\alias{as.dendrogram.CARP}
\alias{as.hclust.CARP}
\title{Print \code{CARP} Results}
\usage{
\method{print}{CARP}(x, ...)

\method{as.dendrogram}{CARP}(object, ...)

\method{as.hclust}{CARP}(x, ...)
}
\arguments{
\item{x}{an object of class \code{CARP} as returned by \code{\link{CARP}}}

\item{...}{Additional unused arguments}

\item{object}{an object of class \code{CARP} as returned by \code{\link{CARP}}}
}
\description{
Prints a brief descripton of a fitted \code{CARP} object.
}
\details{
Reports number of observations and variables of dataset, any preprocessing
done by the \code{\link{CARP}} function, regularization weight information,
and the variant of \code{CARP} used.

The \code{as.dendrogram} and \code{as.hclust} methods convert the
         \code{CBASS} output to an object of class \code{dendrogram} or \code{hclust}
         respectively.
}
\examples{
carp_fit <- CARP(presidential_speech)
print(carp_fit)
}
