% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbass.R
\name{print.CBASS}
\alias{print.CBASS}
\alias{as.dendrogram.CBASS}
\alias{as.hclust.CBASS}
\title{Print \code{CBASS} Results}
\usage{
\method{print}{CBASS}(x, ...)

\method{as.dendrogram}{CBASS}(object, ..., type = c("row", "col"))

\method{as.hclust}{CBASS}(x, ..., type = c("row", "col"))
}
\arguments{
\item{x}{an object of class \code{CARP} as returned by \code{\link{CARP}}}

\item{...}{Additional unused arguments}

\item{object}{an object of class \code{CARP} as returned by \code{\link{CARP}}}

\item{type}{Either \code{"row"} or \code{"col"}, indicating whether
the row or column dendrogram should be returned.}
}
\description{
Prints a brief descripton of a fitted \code{CBASS} object.
}
\details{
The \code{as.dendrogram} and \code{as.hclust} methods convert the
         \code{CBASS} output to an object of class \code{dendrogram} or \code{hclust}
         respectively.
}
\examples{
cbass_fit <- CBASS(X=presidential_speech)
print(cbass_fit)
}
