% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{RBF Kernel Weights}
\alias{RBF Kernel Weights}
\alias{dense_rbf_kernel_weights}
\alias{sparse_rbf_kernel_weights}
\title{Construct Clustering Weights Based on the Radial Basis Function (Gaussian/Euclidean) Kernel}
\usage{
dense_rbf_kernel_weights(
  phi = "auto",
  dist.method = c("euclidean", "maximum", "manhattan", "canberra", "binary",
    "minkowski"),
  p = 2
)

sparse_rbf_kernel_weights(..., k = "auto")
}
\arguments{
\item{phi}{The scale factor used for the RBF kernel}

\item{dist.method}{The type of distance used to calculate distances between
points. See the \code{method} argument to \code{\link[stats]{dist}}.}

\item{p}{The power of the Minkowski distance (only relevant if \code{method == "minkowski"}).
See the \code{p} argument to \code{\link[stats]{dist}}.}

\item{...}{Arguments passed through from \code{sparse_rbf_kernel_weights} to
\code{dense_rbf_kernel_weights}}

\item{k}{The number of neighbors to use}
}
\value{
A function which, when called, returns a matrix of clustering weights.
}
\description{
This is a \emph{factory function} - it returns a \emph{function} which can be
used to create a matrix of clustering weights. In particular, it returns a function
which takes a n-by-p data matrix \eqn{X} and returns an n-by-n matrix whose
\eqn{(i, j)}-th element is given by \eqn{e^{-phi * dist(x_i, x_j)}} where
\eqn{x_i}, \eqn{x_j} are the \eqn{i}-th and \eqn{j}-th row of \eqn{X}
respectively. The distance metric used is determined by the
\code{dist.method} and \code{p} arguments, which are passed to
\code{\link[stats]{dist}}.
}
\details{
The sparse weights (\code{sparse_rbf_kernel_weights}) are calculated by
dropping all but the \code{k} largest weights for each row of the matrix
(equivalent to taking the \code{k} nearest neighbors to each point). The
weight matrix is symmetrized, so if \emph{a} is a neighbor of \emph{b}, but
not vice versa, the edge is still included. If \code{k} is too small, resulting
in a non-fully-connected graph, an error is thrown.

If \code{phi == "auto"}, a grid of possible \eqn{phi} values are used and
the \code{phi} which maximizes the variance of the resulting weights is taken.

If \code{k == "auto"}, the smallest \code{k} that still yields a fully connected
graph is used.
}
\examples{
weight_func <- dense_rbf_kernel_weights()
weight_func(presidential_speech)

weight_func <- dense_rbf_kernel_weights(phi=0.1, dist.method="canberra")
weight_func(presidential_speech)

weight_func <- sparse_rbf_kernel_weights()
weight_func(presidential_speech)
}
