% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveviz.CARP}
\alias{saveviz.CARP}
\title{Save CARP visualizations}
\usage{
\method{saveviz}{CARP}(x, file.name, plot.type = c("path", "dendrogram"),
  image.type = c("dynamic", "static"), static.image.type = c("png"),
  axis = c("PC1", "PC2"), dend.branch.width = 2,
  dend.labels.cex = 0.6, percent = NULL, k = NULL,
  percent.seq = seq(from = 0.05, to = 1, by = 0.05),
  dynamic.width = 1200, dynamic.height = 700, static.width = 8,
  static.height = 5, ...)
}
\arguments{
\item{x}{a CARP object returned by \code{CARP}}

\item{file.name}{name and location of image to be saved.}

\item{plot.type}{a string specifying the type of plot to produce. 'dendrogram'
produces CARP's cluster dendrogram and 'path' produces the
cluster path}

\item{image.type}{a string specifying the type of image output. 'dynamic'
produced a gif of the plot.type along the CARP path. 'static' produces
a single image of plot.type at a point along the CARP path.}

\item{static.image.type}{string specifying the graphics device with which to save
static images.}

\item{axis}{a character vector of length two with elements as 'PC1','PC2',..
etc. Specifics which principal component axis to display for the 'path'
plot.type}

\item{dend.branch.width}{a positive number. Line width on dendrograms.}

\item{dend.labels.cex}{a positive number. Label size on dendrograms.}

\item{percent}{a number between 0 and 1. Specifies how far along the
CARP path the static visualizations should display in terms of
percent reguarlization.}

\item{k}{an interger between 1 and n.obs. Specifies how far along the
CARP path the static visualizations should display in term of
number of clusters.}

\item{percent.seq}{a vector of numbers between 0 and 1 specifying the
positions along the CARP path used to generate dynamic images.}

\item{dynamic.width}{dynamic output width in pixels}

\item{dynamic.height}{dynamic output heigth in pixels}

\item{static.width}{static output width in inches}

\item{static.height}{static output height in inches}

\item{...}{Unused additional generic arguements}
}
\description{
This function will save dynamic CARP visualizations as gifs and static visualizations as image files
}
