% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveviz.CBASS}
\alias{saveviz.CBASS}
\title{Save CBASS visualizations}
\usage{
\method{saveviz}{CBASS}(x, file.name, plot.type = c("heatmap",
  "obs.dendrogram", "var.dendrogram"), image.type = c("dynamic",
  "static"), static.image.type = c("png"), dend.branch.width = 2,
  dend.labels.cex = 0.6, heatrow.label.cex = 1.5,
  heatcol.label.cex = 1.5, percent = NULL, k.obs = NULL,
  k.var = NULL, percent.seq = seq(from = 0.05, to = 1, by = 0.05),
  dynamic.width = 1200, dynamic.height = 700, static.width = 8,
  static.height = 5, ...)
}
\arguments{
\item{x}{a CBASS object returned by \code{CBASS}}

\item{file.name}{name and location of image to be saved.}

\item{plot.type}{a string specifying the type of plot to produce. 'obs.dendrogram'
and 'var.dendrogram' produces the CBASS cluster dendrogram for the observations and
variables respecitvely. 'heatmap' produces a heatmap of the CBASS solution.}

\item{image.type}{a string specifying the type of image output. 'dynamic'
produced a gif of the plot.type along the CARP path. 'static' produces
a single image of plot.type at a point along the CARP path.}

\item{static.image.type}{string specifying the graphics device with which to save
static images.}

\item{dend.branch.width}{a positive number. Line width on dendrograms.}

\item{dend.labels.cex}{a positive number. Label size on dendrograms.}

\item{heatrow.label.cex}{heatmap row label size}

\item{heatcol.label.cex}{heatmap column label size}

\item{percent}{a number between 0 and 1. Specifies how far along the
CARP path the static visualizations should display via the amount of
regularization.}

\item{k.obs}{An interger between 1 and \code{n.obs}.
Specifies how far along the
CARP path the static visualizations should display via the number of unique
observation clusters.}

\item{k.var}{An interger between 1 and \code{p.var}.
Specifies how far along the
CARP path the static visualizations should display via the number of unique
variable clusters.}

\item{percent.seq}{a vector of numbers between 0 and 1 specifying the
positions along the CARP path used to generate dynamic images.}

\item{dynamic.width}{dynamic output width in pixels}

\item{dynamic.height}{dynamic output heigth in pixels}

\item{static.width}{static output width in inches}

\item{static.height}{static output height in inches}

\item{...}{Unused additional generic arguements}
}
\description{
This function will save dynamic CBASS visualizations as gifs and static visualizations as image files
}
