% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveviz}
\alias{saveviz}
\title{Method for saving CARP and CBASS visualizations}
\usage{
saveviz(x, ...)
}
\arguments{
\item{x}{a CARP or CBASS object}

\item{...}{additional arguements passed to \code{saveviz.CARP} or
\code{saveviz.CBASS}}
}
\description{
See \code{saveviz.CARP} and \code{saveviz.CBASS} for details
}
