% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SparseWeights}
\alias{SparseWeights}
\title{Compute sparse kNN weight vector from dense weights}
\usage{
SparseWeights(X, dense.weights, k)
}
\arguments{
\item{X}{an n by p matrix with rows the observations and columns the variables.}

\item{dense.weights}{a vector dense weights, such as those returned
by \code{DenseWeights}}

\item{k}{a positive integer less than nobs. The number of nearest neighbors.}
}
\value{
sparse.weights a numeric vector of sparse weights
}
\description{
Compute sparse kNN weight vector from dense weights
}
\keyword{internal}
