% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SparseWeights_LEGACY}
\alias{SparseWeights_LEGACY}
\title{(slowly) Compute sparse kNN weight vector from dense weights}
\usage{
SparseWeights_LEGACY(X, dense.weights, k)
}
\arguments{
\item{X}{an n by p matrix with rows the observations and columns the variables.}

\item{dense.weights}{a vector dense weights, such as those returned
by \code{DenseWeights}}

\item{k}{a positive integer less than nobs. The number of nearest neighbors.}
}
\value{
sparse.weights a numeric vector of sparse weights
}
\description{
(slowly) Compute sparse kNN weight vector from dense weights
}
\keyword{internal}
