% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_breast}
\alias{tcga_breast}
\title{Log-Transformed Level III RPKM Gene Expression Levels for 438 Breast-Cancer Patients}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 438 rows and 353 columns.
}
\usage{
tcga_breast
}
\description{
This data set (\eqn{n = 438, p = 353}) contains log-transformed Level III RPKM gene 
expression levels for 438 breast-cancer patients collected by the Cancer Genome Atlas
Network. The Luminal A and Luminal B subtypes have been combined. The row names give the
clinically diagnosed subtype (true cluster labels) and the column names are the gene IDs.
}
\references{
The Cancer Genome Atlas Network. "Comprehensive Molecular Portraits of Human Breast Tumours"
Nature 490, p.61-70. 2012. \doi{10.1038/nature11412}
}
\keyword{datasets}
