% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WeightAdjacency}
\alias{WeightAdjacency}
\title{Construct adjacency matrix induced by weights}
\usage{
WeightAdjacency(weights, nobs)
}
\arguments{
\item{weights}{a vector of weights such as returned by \code{SparseWeights}}

\item{nobs}{the number of observations being clustered}
}
\value{
adj a sparse adjacency matrix
}
\description{
Construct adjacency matrix induced by weights
}
