% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carp.R
\name{carp.control}
\alias{carp.control}
\title{Control for \code{CARP} fits}
\usage{
carp.control(obs.labels = NULL, var.labels = NULL, X.center = TRUE,
  X.scale = FALSE, phi = NULL, rho = 1, weights = NULL, k = NULL,
  weight.dist = "euclidean", weight.dist.p = 2, ncores = 1L,
  max.iter = 1000000L, burn.in = 50L, alg.type = "carpviz",
  t = 1.05, npcs = NULL, dendrogram.scale = NULL, ...)
}
\arguments{
\item{obs.labels}{A character vector of length \eqn{n}: observations (row) labels}

\item{var.labels}{A character vector of length \eqn{p}: variable (column) labels}

\item{X.center}{A logical: Should \code{X} be centered columnwise?}

\item{X.scale}{A logical: Should \code{X} be scaled columnwise?}

\item{phi}{A positive real number: the scale factor used in the RBF kernel}

\item{rho}{For advanced users only (not advisable to change): the penalty
parameter used for the augmented Lagrangian.}

\item{weights}{A vector of positive number of length \code{choose(n,2)}.}

\item{k}{An positive integer: the number of neighbors used to create sparse weights}

\item{weight.dist}{A string indicating the distance metric used to calculate weights.
See \code{\link[stats]{distance}} for details.}

\item{weight.dist.p}{The exponent used to calculate the Minkowski distance if
\code{weight.dist = "minkowski"}.
See \code{\link[stats]{distance}} for details.}

\item{ncores}{An positive integer: the number of cores to use.}

\item{max.iter}{An integer: the maximum number of CARP iterations.}

\item{burn.in}{An integer: the number of initial iterations at a fixed
(small) value of \eqn{\lambda}}

\item{alg.type}{Which \code{CARP} variant to use. Allowed values are \itemize{
\item \code{"carp"} - The standard \code{CARP} algorithm with \eqn{L2} penalty;
\item \code{"carpviz"} - The back-tracking \code{CARP} algorithm with \eqn{L2} penalty;
\item \code{"carpl1"} - The standard \code{CARP} algorithm with \eqn{L1} penalty; and
\item \code{"carpvizl1"} - The back-tracking \code{CARP} algorithm with \eqn{L1} penalty.}}

\item{t}{A number greater than 1: the size of the multiplicative update to
the cluster fusion regularization parameter (not used by
back-tracking variants). Typically on the scale of \code{1.005} to \code{1.1}.}

\item{npcs}{An integer >= 2. The number of principal components to compute
for path visualization.}

\item{dendrogram.scale}{A character string denoting how the scale of dendrogram
regularization proportions should be visualized. Choices are \code{'original'}
or \code{'log'}; if not provided, a data-driven heuristic choice is used.}

\item{...}{Unused arguements. An error will be thrown if any unrecognized
arguments as given.}
}
\value{
A list containing the \code{CARP} algorithm parameters.
}
\description{
Set \code{CARP} algorithm parameters
}
\details{
This function constructs a list containing additional arguments to control
the behavior of the \code{CARP} algorithm. It is typically only used internally
by \code{\link{CARP}}, but may be useful to advanced users who wish to
construct the \code{control} argument directly.
}
