% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CARP}
\alias{CARP}
\title{Compute CARP solution path}
\usage{
CARP(X, verbose = 1, interactive = TRUE, static = TRUE, control = NULL,
  ...)
}
\arguments{
\item{X}{A n by p matrix with rows the observations and columns the variables}

\item{verbose}{either 0,1,or 2. Higher numbers indicate more verbosity while
computing.}

\item{interactive}{A logical. Should interactive paths and dendrograms be
returned?}

\item{static}{A logical. Should static paths and dendrograms be returned?}

\item{control}{A list. list of parameters for CARP fitting. see carp.control.}

\item{...}{arguements to carp.control}
}
\value{
X the original data matrix

carp.dend if static==TRUE, a dendrogam representation of the clustering solution path

carp.cluster.path.vis the CARP solution path

n.obs the number of observations

p.var the number of variables

phi A positive numner used for scaling in RBF kernel

k an integer >= 1. The number of neighbors used to create sparse weights

burn.in an integer. The number of initial iterations at a fixed
value of (small) lambda_k

alg.type Which CARP algorithm to perform. Choices are 'carpviz'
and 'carp;

X.center A logical. Should X be centered?

X.scale A logical. Should X be scaled?
}
\description{
\code{CARP} returns a fast approximation to the Convex Clustering
solution path along with visualizations such as dendrograms and
cluster paths. Visualizations may be static, interactive, or both.
}
\details{
\code{CARP} solves the Convex Clustering problem via
Algorithmic Regularization Paths. A seqeunce of clustering
solutions is returned along with several visualizations.
}
\examples{
library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech$X[1:10,1:4]
carp.fit <- CARP(
    X=Xdat,
    obs.labels=presidential_speech$labels[1:10])
carp.fit
}
