% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carp.R
\name{CARP}
\alias{CARP}
\title{Compute \code{CARP} (Convex Clustering) Solution Path}
\usage{
CARP(X, verbose = 1L, interactive = TRUE, static = TRUE, ...,
  control = NULL)
}
\arguments{
\item{X}{The data matrix (\eqn{X \in R^{n \times p}}{X}): rows correspond to
the observations (to be clustered) and columns to the variables (which
will not be clustered).}

\item{verbose}{Any of the values \code{0}, \code{1}, or \code{2}. Higher values
correspond to more verbose output while running.}

\item{interactive}{A logical. Should interactive paths and dendrograms be
returned?}

\item{static}{A logical. Should static paths and dendrograms be returned?}

\item{...}{Additional arguments used to control the behavior of \code{CARP}; see
\code{\link{carp.control}} for details.}

\item{control}{A list containing advanced parameters for the \code{CARP} algorithm,
typically created by \code{\link{carp.control}}.}
}
\value{
An object of class \code{CARP} containing the following elements (among others):
        \itemize{
        \item \code{X}: the original data matrix
        \item \code{n.obs}: the number of observations (rows of \code{X})
        \item \code{p.var}: the number of variables (columns of \code{X})
        \item \code{alg.type}: the \code{CARP} variant used
        \item \code{X.center}: a logical indicating whether \code{X} was centered
                               column-wise before clustering
        \item \code{X.scale}: a logical indicating whether \code{X} was scaled
                              column-wise before centering
        \item \code{burn.in}: an integer indicating the number of "burn-in"
                              iterations performed
        \item \code{k}: the number of neighbors used to create sparse clustering weights
        \item \code{phi}: the scale factor of the RBF kernel used to calculate
                          clustering weights
        \item \code{carp.dend}: If \code{static=TRUE}, an dendrogram (object of
                                class \code{\link[stats]{hclust}}) containing
                                the clustering solution path
        \item \code{carp.cluster.path.vis}: The \code{CARP} solution path
        \item \code{static}: a logical indicating whether static visualizations are available for this \code{CARP} object
        \item \code{interactive}: a logical indicating whether interactive visualizations are available for this \code{CARP} object
        }
}
\description{
\code{CARP} returns a fast approximation to the Convex Clustering
solution path along with visualizations such as dendrograms and
cluster paths. Visualizations may be static, interactive, or both.
}
\details{
\code{CARP} solves the Convex Clustering problem via
Algorithmic Regularization Paths. A seqeunce of clustering
solutions is returned along with several visualizations.
}
\examples{
carp_fit <- CARP(presidential_speech[1:10,1:4])
print(carp_fit)
plot(carp_fit)
}
