% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbass.R
\name{cbass.control}
\alias{cbass.control}
\title{Control for \code{CBASS} fits}
\usage{
cbass.control(obs.labels = NULL, var.labels = NULL,
  X.center.global = TRUE, rho = 1, phi = NULL, weights.obs = NULL,
  weights.var = NULL, obs.weight.dist = "euclidean",
  obs.weight.dist.p = 2, var.weight.dist = "euclidean",
  var.weight.dist.p = 2, k.obs = NULL, k.var = NULL, t = 1.01,
  ncores = as.integer(1), max.iter = as.integer(1e+06),
  burn.in = as.integer(50), alg.type = "cbassviz",
  npcs = as.integer(4), ...)
}
\arguments{
\item{obs.labels}{A character vector of length \eqn{n}: observations (row) labels}

\item{var.labels}{A character vector of length \eqn{p}: variable (column) labels}

\item{X.center.global}{A logical: Should \code{X} be centered globally?
\emph{I.e.}, should the global mean of \code{X} be subtracted?}

\item{rho}{For advanced users only (not advisable to change): the penalty
parameter used for the augmented Lagrangian.}

\item{phi}{A positive real number: the scale factor used in the RBF kernel}

\item{weights.obs}{A vector of positive number of length \code{choose(n,2)}.}

\item{weights.var}{A vector of positive number of length \code{choose(n,2)}.}

\item{obs.weight.dist}{A string indicating the distance metric used to calculate weights.
See \code{\link[stats]{distance}} for details.}

\item{obs.weight.dist.p}{The exponent used to calculate the Minkowski distance if
\code{weight.dist = "minkowski"}.
See \code{\link[stats]{distance}} for details.}

\item{var.weight.dist}{A string indicating the distance metric used to calculate weights.
See \code{\link[stats]{distance}} for details.}

\item{var.weight.dist.p}{The exponent used to calculate the Minkowski distance if
\code{weight.dist = "minkowski"}.
See \code{\link[stats]{distance}} for details.}

\item{k.obs}{An positive integer: the number of neighbors used to create sparse weights}

\item{k.var}{An positive integer: the number of neighbors used to create sparse weights}

\item{t}{A number greater than 1: the size of the multiplicative update to
the cluster fusion regularization parameter (not used by
back-tracking variants). Typically on the scale of \code{1.005} to \code{1.1}.}

\item{ncores}{An positive integer: the number of cores to use.}

\item{max.iter}{An integer: the maximum number of CARP iterations.}

\item{burn.in}{An integer: the number of initial iterations at a fixed
(small) value of \eqn{\lambda}}

\item{alg.type}{Which \code{CBASS} variant to use. Allowed values are \itemize{
\item \code{"cbass"} - The standard \code{CBASS} algorithm with \eqn{L2} penalty;
\item \code{"cbassviz"} - The back-tracking \code{CBASS} algorithm with \eqn{L2} penalty;
\item \code{"cbassl1"} - The standard \code{CBASS} algorithm with \eqn{L1} penalty; and
\item \code{"cbassvizl1"} - The back-tracking \code{CBASS} algorithm with \eqn{L1} penalty.}}

\item{npcs}{An integer >= 2. The number of principal components to compute
for path visualization.}

\item{...}{Unused arguements. An error will be thrown if any unrecognized
arguments as given.}
}
\value{
A list containing the \code{CBASS} algorithm parameters.
}
\description{
Set \code{CBASS} algorithm parameters
}
\details{
This function constructs a list containing additional arguments to control
the behavior of the \code{CBASS} algorithm. It is typically only used internally
by \code{\link{CBASS}}, but may be useful to advanced users who wish to
construct the \code{control} argument directly.
}
