% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbass.R
\name{CBASS}
\alias{CBASS}
\title{Compute \code{CBASS} (Convex BiClustering) Solution Path}
\usage{
CBASS(
  X,
  ...,
  row_weights = sparse_rbf_kernel_weights(k = "auto", phi = "auto", dist.method =
    "euclidean", p = 2),
  col_weights = sparse_rbf_kernel_weights(k = "auto", phi = "auto", dist.method =
    "euclidean", p = 2),
  row_labels = rownames(X),
  col_labels = colnames(X),
  X.center.global = TRUE,
  t = 1.01,
  back_track = FALSE,
  exact = FALSE,
  norm = 2,
  npcs = min(4L, NCOL(X), NROW(X)),
  dendrogram.scale = NULL,
  status = (interactive() && (clustRviz_logger_level() \%in\% c("MESSAGE", "WARNING",
    "ERROR")))
)
}
\arguments{
\item{X}{The data matrix (\eqn{X \in R^{n \times p}}{X}).
If \code{X} has missing values - \code{NA} or
\code{NaN} values - they will be automatically imputed.}

\item{...}{Unused arguements. An error will be thrown if any unrecognized
arguments as given.}

\item{row_weights}{One of the following: \itemize{
\item A function which, when called with argument \code{X},
      returns a n-by-n matrix of fusion weights.
\item A matrix of size n-by-n containing fusion weights
}
Note that the weights will be renormalized to sum to
\eqn{1/\sqrt{n}} internally.}

\item{col_weights}{One of the following: \itemize{
\item A function which, when called with argument \code{t(X)},
      returns a p-by-p matrix of fusion weights. (Note the
      transpose.)
\item A matrix of size p-by-p containing fusion weights
}
Note that the weights will be renormalized to sum to
\eqn{1/\sqrt{p}} internally.}

\item{row_labels}{A character vector of length \eqn{n}: row (observation) labels}

\item{col_labels}{A character vector of length \eqn{p}: column (variable) labels}

\item{X.center.global}{A logical: Should \code{X} be centered globally?
\emph{I.e.}, should the global mean of \code{X} be subtracted?}

\item{t}{A number greater than 1: the size of the multiplicative update to
the cluster fusion regularization parameter (not used by
back-tracking variants). Typically on the scale of \code{1.005} to \code{1.1}.}

\item{back_track}{A logical: Should back-tracking be used to exactly identify fusions?
By default, back-tracking is not used.}

\item{exact}{A logical: Should the exact solution be computed using an iterative algorithm?
By default, algorithmic regularization is applied and the exact solution
is not computed. Setting \code{exact = TRUE} often significantly increases
computation time.}

\item{norm}{Which norm to use in the fusion penalty? Currently only \code{1}
and \code{2} (default) are supported.}

\item{npcs}{An integer >= 2. The number of principal components to compute
for path visualization.}

\item{dendrogram.scale}{A character string denoting how the scale of dendrogram
regularization proportions should be visualized.
Choices are \code{'original'} or \code{'log'}; if not
provided, a data-driven heuristic choice is used.}

\item{status}{Should a status message be printed to the console?}
}
\value{
An object of class \code{CBASS} containing the following elements (among others):
        \itemize{
        \item \code{X}: the original data matrix
        \item \code{n}: the number of observations (rows of \code{X})
        \item \code{p}: the number of variables (columns of \code{X})
        \item \code{alg.type}: the \code{CBASS} variant used
        \item \code{row_fusions}: A record of row fusions - see the documentation
                                  of \code{\link{CARP}} for details of what this
                                  may include.
        \item \code{col_fusions}: A record of column fusions - see the documentation
                                  of \code{\link{CARP}} for details of what this
                                  may include.
        }
}
\description{
\code{CBASS} returns a fast approximation to the Convex BiClustering
solution path along with visualizations such as dendrograms and
heatmaps. \code{CBASS} solves the Convex Biclustering problem using an efficient
Algorithmic Regularization scheme.
}
\examples{
\dontrun{
cbass_fit <- CBASS(presidential_speech)
print(cbass_fit)
plot(cbass_fit)
}
}
