% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CBASS}
\alias{CBASS}
\title{Compute CBASS solution path}
\usage{
CBASS(X, obs.labels = NULL, var.labels = NULL, X.center.global = TRUE,
  rho = 1, phi = 0.1, weights.obs = NULL, weights.vars = NULL,
  k.obs = NULL, k.var = NULL, t = NULL, ncores = 1, max.iter = 1e+06,
  burn.in = 50, verbose = 1, alg.type = "cbassviz", interactive = TRUE,
  static = TRUE)
}
\arguments{
\item{X}{A n.obs by p.vars matrix with rows the observations and columns the variables}

\item{obs.labels}{a vector of length n.obs containing observations (row) labels}

\item{var.labels}{a vector of length p.vars containing variable (column) labels}

\item{X.center.global}{a logical. If TRUE, the global mean of X is removed.}

\item{rho}{A positive number for augmented lagrangian. Not advisable to change.}

\item{phi}{A positive numner used for scaling in RBF kernel}

\item{weights.obs}{A vector of positive number of length choose(n.obs,2).
Determines observation pair fusions weight.}

\item{weights.vars}{A vector of positive number of length choose(p.vars,2).
Determines variable pair fusions weight.}

\item{k.obs}{an integer >= 1. The number of neighbors used to create sparse
observation weights}

\item{k.var}{an integer >= 1. The number of neighbors used to create sparse
variable weights}

\item{t}{a number greater than 1. The size of the multiplicitive
regularization parameter update. Typical values are: 1.1, 1.05, 1.01, 1.005.
Not used CBASS-VIZ algorithms.}

\item{ncores}{an integer >= 1. The number of cores to use.}

\item{max.iter}{an integer. The maximum number of CARP iterations.}

\item{burn.in}{an integer. The number of initial iterations at a fixed
value of (small) lambda_k}

\item{verbose}{either 0,1,or 2. Higher numbers indicate more verbosity while
computing.}

\item{alg.type}{Which CARP algorithm to perform. Choices are 'cbassviz'
and 'cbass';}

\item{interactive}{A logical. Should an interactive heatmap be returned?}

\item{static}{A logical. Should observation and variable dendrograms
be returned?}
}
\value{
X the original data matrix

cbass.cluster.path.obs the CBASS observation solution path

cbass.cluster.path.var the CBASS variable solution path

cbass.dend.obs if static==TRUE, a dendrogam representation of the
observation clustering solution path

cbass.dend.var if static==TRUE, a dendrogam representation of the
variable clustering solution path

phi.obs A positive numner used for scaling in observation RBF kernel

phi.var A positive numner used for scaling in variable RBF kernel

k.obs an integer >= 1. The number of neighbors used to create sparse
observation weights

k.var an integer >= 1. The number of neighbors used to create sparse
variable weights

burn.in an integer. The number of initial iterations at a fixed
value of (small) lambda_k

alg.type Which CARP algorithm to perform. Choices are 'cbassviz'
and 'cbass'

interactive A logical. Should an interactive heatmap be returned?

static A logical. Should observation and variable dendrograms
be returned?

obs.labels a vector of length n.obs containing observations (row) labels

var.labels a vector of length p.vars containing variable (column) labels

X.center.global a logical. If TRUE, the global mean of X is removed.
}
\description{
\code{CBASS} returns a fast approximation to the Convex BiClustering
solution path along with visualizations such as dendrograms and
heatmaps. Visualizations may be static, interactive, or both.
}
\details{
\code{CBASS} solves the following problem.
}
