% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbass.R
\name{CBASS}
\alias{CBASS}
\title{Compute \code{CBASS} (Convex BiClustering) Solution Path}
\usage{
CBASS(X, verbose = 1L, interactive = TRUE, static = TRUE, ...,
  control = NULL)
}
\arguments{
\item{X}{The data matrix (\eqn{X \in R^{n \times p}}{X}): rows correspond to
the observations (to be clustered) and columns to the variables (which
will not be clustered).}

\item{verbose}{Any of the values \code{0}, \code{1}, or \code{2}. Higher values
correspond to more verbose output while running.}

\item{interactive}{A logical. Should interactive paths and dendrograms be
returned?}

\item{static}{A logical. Should static paths and dendrograms be returned?}

\item{...}{Additional arguments used to control the behavior of \code{CBASS}; see
\code{\link{cbass.control}} for details.}

\item{control}{A list containing advanced parameters for the \code{CBASS} algorithm,
typically created by \code{\link{cbass.control}}.}
}
\value{
An object of class \code{CBASS} containing the following elements (among others):
        \itemize{
        \item \code{X}: the original data matrix
        \item \code{n.obs}: the number of observations (rows of \code{X})
        \item \code{p.var}: the number of variables (columns of \code{X})
        \item \code{alg.type}: the \code{CBASS} variant used
        \item \code{X.center.global}: a logical indicating whether \code{X}
                                      was globally centered before clustering
        \item \code{X.scale}: a logical indicating whether \code{X} was scaled
                              column-wise before centering
        \item \code{burn.in}: an integer indicating the number of "burn-in"
                              iterations performed
        \item \code{k.obs}: the number of neighbors used to create sparse
                            clustering weights for the observations
        \item \code{k.var}: the number of neighbors used to create sparse
                            clustering weights for the variables
        \item \code{phi.obs}: the scale factor of the RBF kernel used to calculate
                              clustering weights for the observations
        \item \code{phi.var}: the scale factor of the RBF kernel used to calculate
                              clustering weights for the variables
        \item \code{carp.dend.obs}: If \code{static=TRUE}, an dendrogram (object of
                                    class \code{\link[stats]{hclust}}) containing
                                    the clustering solution path for the observations
        \item \code{carp.dend.var}: If \code{static=TRUE}, an dendrogram (object of
                                    class \code{\link[stats]{hclust}}) containing
                                    the clustering solution path for the variables
        \item \code{cbass.cluster.path.obs}: The \code{CBASS} solution path for the observations
        \item \code{cbass.cluster.path.var}: The \code{CBASS} solution path for the variables
        \item \code{static}: a logical indicating whether static visualizations are available for this \code{CBASS} object
        \item \code{interactive}: a logical indicating whether interactive visualizations are available for this \code{CBASS} object
        \item \code{obs.labels}: a character vector of length \code{n.obs} containing
                                 observation (row) labels
        \item \code{var.labels}: a character vector of length \code{p.var} containing
                                 variable (column) labels
        }

obs.labels a vector of length n.obs containing observations (row) labels

var.labels a vector of length p.var containing variable (column) labels
}
\description{
\code{CBASS} returns a fast approximation to the Convex BiClustering
solution path along with visualizations such as dendrograms and
heatmaps. Visualizations may be static, interactive, or both.
}
\details{
\code{CBASS} solves the Convex Biclustering problem via
Algorithmic Regularization Paths. A seqeunce of biclustering
solutions is returned along with several visualizations.
}
\examples{
\dontrun{
cbass_fit <- CBASS(presidential_speech)
print(cbass_fit)
plot(cbass_fit)
}
}
