% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{clustering.CBASS}
\alias{clustering.CBASS}
\title{Get biclustering solution from a CBASS object}
\usage{
\method{clustering}{CBASS}(x, k.obs = NULL, k.var = NULL,
  percent = NULL, ...)
}
\arguments{
\item{x}{A CBASS object returned by \code{CBASS}}

\item{k.obs}{An interger between 1 and \code{n.obs}. The number of unique
observation clusters.}

\item{k.var}{An interger between 1 and \code{p.var}. The number of unique
variable clusters.}

\item{percent}{A number between 0 and 1. The percent of regularization at
which to cut the path.}

\item{...}{Unused additional generic arguements}
}
\value{
A list with elements
\describe{
\item{\code{clustering.assignment.obs}}{
A vector of observation cluster labels of length \code{n.obs}.
}
\item{\code{clustering.assignment.var}}{
A vector of variable cluster labels of length \code{p.var}.
}
\item{\code{cluster.mean.matrix}}{
The cluster mean matrix of size \code{n.obs} by \code{p.var}.
}
}
}
\description{
Returns (observation and variable) cluster labels and
cluster mean matrix at a point along the CBASS path.
}
\details{
Passing either the desired number of clusters (\code{k}) or the percent
regularization (\code{percent}) returns the clustering assignment
and cluster mean matrix at the specific point along the CBASS path.
}
\examples{
\dontrun{
library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech
cbass.fit <- CBASS(X=Xdat)
cbass.clustering <- clustering(cbass.fit,percent = .8)
}
}
