% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{clustering}
\alias{clustering}
\title{Method for returning CARP and CBASS clustering solutions}
\usage{
clustering(x, ...)
}
\arguments{
\item{x}{a CARP or CBASS object}

\item{...}{additional arguements passed to \code{clustering.CARP} or
\code{clustering.CBASS}}
}
\value{
CARP clustering solutions or CBASS biclustering solutions
}
\description{
See \code{clustering.CARP} and \code{clustering.CBASS} for details
}
