% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DenseWeights}
\alias{DenseWeights}
\title{Compute dense distanced-based gaussian kernel weights for use in CARP or CBASS}
\usage{
DenseWeights(X, phi = 1, method = c("euclidean", "maximum",
  "manhattan", "canberra", "binary", "minkowski"), p = 2)
}
\arguments{
\item{X}{an n by p matrix with rows the observations and columns the variables.}

\item{phi}{a number. Scaling parameter used in the gaussian kernel.}

\item{method}{a string. Passed to the \code{dist} function. See \code{?dist}}

\item{p}{The power of Minkowski distance. Passed to the \code{dist} function. See \code{?dist}}
}
\value{
dense.weights a numeric vector of weights
}
\description{
Compute dense distanced-based gaussian kernel weights for use in CARP or CBASS
}
\examples{
\dontrun{
library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech$X
Xdat.preprocessed <- scale(Xdat,center=TRUE,scale=FALSE)
dense.weights <- DenseWeights(X = Xdat.preprocessed,phi=1e-3)
}
}
\keyword{internal}
