% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MinKNN}
\alias{MinKNN}
\title{Compute the minimum number of nearest neighbors required to fully
cluster all observations}
\usage{
MinKNN(X, dense.weights)
}
\arguments{
\item{X}{an n.obs x p.vars matrix}

\item{dense.weights}{a vector vector of dense weights such as returned by
\code{DenseWeights}}
}
\value{
k an integer. The smallest number of nearest neighbors required to
fully cluster all observations.
}
\description{
Compute the minimum number of nearest neighbors required to fully
cluster all observations
}
\examples{
\dontrun{
library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech$X
Xdat.preprocessed <- scale(Xdat,center=TRUE,scale=FALSE)
dense.weights <- DenseWeights(X = Xdat.preprocessed,phi=1e-3)
MinKNN(
X = Xdat.preprocessed,
dense.weights = dense.weights
) -> k.min
}
}
