% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cbass.R
\name{plot.CBASS}
\alias{plot.CBASS}
\title{Plot function for CBASS objects}
\usage{
\method{plot}{CBASS}(x, type = c("obs.dendrogram", "var.dendrogram",
  "heatmap", "interactive"), dend.branch.width = 2,
  dend.labels.cex = 0.6, heatrow.label.cex = 1.5,
  heatcol.label.cex = 1.5, ...)
}
\arguments{
\item{x}{a CBASS object returned by \code{CARP}}

\item{type}{a string specifying the type of plot to produce.
'obs.dendrogram' produces the static cluster dendrogram for the observations;
'var.dendrogram' produces the static cluster dendrogram for the variables;
'heatmap' produces the static cluster heatmap with assiociated observation
and variable dendrograms;
and 'interactive' produces an interactive visualization of cluster
heatmap and its associated dendrograms.}

\item{dend.branch.width}{a positive number. Line width on dendrograms.}

\item{dend.labels.cex}{a positive number. Label size on dendrograms.}

\item{heatrow.label.cex}{heatmap row label size}

\item{heatcol.label.cex}{heatmap column label size}

\item{...}{Unused additional generic arguements}
}
\description{
\code{plot.CBASS} displays both static and interactive visualizations of
the CBASS clustering solution path, including dendrograms and heatmaps.
}
\details{
Possible visualizations of the CBASS solution path are: (i) a static
dendrogram representing the clustering solution path at various
levels of regularization (both observations and variables); and (ii) an
interactive cluster heatmap, showing
how the clustering solutions change with increased reguarliazation.
}
\examples{
\dontrun{
library(clustRviz)
data("presidential_speech")
Xdat <- presidential_speech[1:10,1:4]
cbass.fit <- CARP(X=Xdat)
plot(cbass.fit,type='interactive')
}
}
