% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cbass.R, R/save.R
\name{plot.CBASS}
\alias{plot.CBASS}
\alias{saveviz.CBASS}
\title{Visualize the results of Convex BiClustering (\code{CBASS})}
\usage{
\method{plot}{CBASS}(
  x,
  ...,
  type = c("heatmap", "row.dendrogram", "col.dendrogram", "row.path", "col.path", "js",
    "interactive"),
  percent,
  k.row,
  k.col,
  dend.branch.width = 2,
  dend.labels.cex = 0.6,
  dend.ylab.cex = 1.2,
  heatrow.label.cex = 1,
  heatcol.label.cex = 1,
  margins = c(5, 5),
  axis = c("PC1", "PC2")
)

\method{saveviz}{CBASS}(
  x,
  file.name,
  type = c("heatmap", "row.dendrogram", "col.dendrogram", "js"),
  dynamic = TRUE,
  dend.branch.width = 2,
  dend.labels.cex = 0.6,
  heatrow.label.cex = 1.5,
  heatcol.label.cex = 1.5,
  percent,
  k.row,
  k.col,
  percent.seq = seq(from = 0, to = 1, by = 0.01),
  width = 8,
  height = 5,
  margins = c(5, 5),
  units = c("in", "cm", "mm", "px"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{CBASS} as returned by \code{\link{CBASS}}}

\item{...}{Additional arguments, which are handled differently for different
values of \code{type}.\itemize{
\item When \code{type} is \code{"heatmap"}, \code{"row.path"},
      \code{"col.path"}, or \code{"interactive"}, the presence of
      unknown arguments triggers an error;
\item when \code{type == "row.dendrogram"} or \code{type == "col.dendrogram"},
      \code{...} is forwarded to \code{\link[stats]{plot.dendrogram}}; and
\item when \code{type == "js"}, \code{...} is forwarded to
      \code{\link[heatmaply]{heatmaply}}.
} See the documentation of the linked functions for details about
additional supported arguments. \code{saveviz} passes arguments
to the corresponding plot \code{type}.}

\item{type}{A string indicating the type of visualization to show (see details above).}

\item{percent}{A number between 0 and 1, giving the regularization level (as
a fraction of the final regularization level used) at which to
assign clusters in the static (\code{type = "dendrogram"} or \code{type = "path"})
plots.}

\item{k.row}{An integer indicating the desired number of row clusters to be displayed
in the static plots. (If plotting columns, the regularization level
giving \code{k.row} rows clusters is used.)}

\item{k.col}{An integer indicating the desired number of column clusters to be displayed
in the static plots. (If plotting rows, the regularization level
giving \code{k.col} column clusters is used.)}

\item{dend.branch.width}{Branch width for dendrogram plots (ignored for
other plot types) - must be positive.}

\item{dend.labels.cex}{Label size for dendrogram plots (ignored for other plot
types) - must be positive.}

\item{dend.ylab.cex}{Y-axis size for dendrogram plots (ignored for other plot
types) - must be positive.}

\item{heatrow.label.cex}{heatmap row label size}

\item{heatcol.label.cex}{heatmap column label size}

\item{margins}{A vector of length 2 specifying margin sizes. See the \code{margin}
argument to \code{\link[gplots]{heatmap.2}}.}

\item{axis}{A character vector of length two indicating which features or principal
components to use as the axes in the path visualizations.
Currently only features like \code{"PC1"} or \code{"PC2"} (indicating
the first principal component projections) are supported.}

\item{file.name}{The name of the output file. The type of resulting image
is determined from the extension.If \code{dynamic = TRUE},
the extension is changed to \code{.gif} internally. If \code{type = "js"},
the extension is changed to \code{.html} internally.}

\item{dynamic}{Should the resulting animation be dynamic (animated) or not?
If \code{TRUE}, a dynamic visualization which varies along the
\code{CARP} solution path at a grid given by \code{percent.seq}
is produced (as a \code{GIF}). If \code{FALSE}, a fixed visualization
at a single solution (determined by \code{percent}, \code{k.row} or
\code{k.col} if supplied) is produced.
Currently ignored when \code{type = "js"}.}

\item{percent.seq}{A grid of values of \code{percent} along which to generate
dynamic visualizations (if \code{dynamic == TRUE})}

\item{width}{The width of the output, given in \code{unit}s}

\item{height}{The height of the output, given in \code{unit}s}

\item{units}{The unit in which \code{width} and \code{height} are specified}
}
\value{
The value of the return type depends on the \code{type} argument:\itemize{
 \item if \code{type \%in\% c("row.dendrogram", "col.dendrogram", "heatmap")},
        \code{x} is returned invisibly;
 \item if \code{type \%in\% c("row.path", "col.path")}, an object of class
       \code{\link[ggplot2]{ggplot}} which can be plotted directly (by invoking
       its print method) or modified further by the user is returned;
 \item if \code{type = "interactive"}, a \code{\link[shiny]{shiny}} app which can be activated
       by invoking its print method.
} \code{saveviz.CBASS} always returns \code{file.name} invisibly.
}
\description{
\code{plot.CBASS} provides a range of ways to visualize the results of convex
clustering, including: \itemize{
\item Dendrograms, illustrating the nested cluster hierarchy inferred from
      the convex clustering solution path (\code{type = "row.dendrogram"} and
      \code{type = "col.dendrogram"} for the row (observation) and column
      (feature / variable) clusterings, respectively);
\item Path plots, showing the coalescence of the estimated cluster centroids
      as the regularization parameter is increased (\code{type = "row.dendrogram"}
      and \code{type = "col.dendrogram"} for the row (observation) and column
      (feature / variable) clusterings, respectively);
\item A cluster heatmap, displaying row and column histograms, as well
      as the clustered data matrix in a single graphic (\code{type = "heatmap"});
\item An interactive Javascript cluster heatmap (\code{type = "heatmap"}); and
\item A \code{\link[shiny]{shiny}} app, which can display the clustering solutions
      as a "movie" or allow for interactive exploration (\code{type = "interactive"}).
}
}
\details{
The \code{\link{saveviz.CBASS}} function provides a unified interface
         for exporting \code{CBASS} visualizations to files. For all plots,
         at most one of \code{percent}, \code{k.row}, and \code{k.col} must be supplied.
}
\examples{
cbass_fit <- CBASS(presidential_speech)
plot(cbass_fit)
plot(cbass_fit, type = "heatmap")
\dontrun{
plot(cbass_fit, type='interactive')
}
}
