% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carp.R
\name{print.CARP}
\alias{print.CARP}
\title{Print \code{CARP} Results}
\usage{
\method{print}{CARP}(x, ...)
}
\arguments{
\item{x}{an object of class \code{CARP} as returned by \code{\link{CARP}}}

\item{...}{Additional unused arguments}
}
\description{
Prints a brief descripton of a fitted \code{CARP} object.
}
\details{
Reports number of observations and variables of dataset, any preprocessing
done by the \code{\link{CARP}} function, regularization weight information,
the variant of \code{CARP} used, and the visualizations returned.
}
\examples{
carp_fit <- CARP(presidential_speech[1:10,1:4])
print(carp_fit)
}
