% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbass.R
\name{print.CBASS}
\alias{print.CBASS}
\title{Print \code{CBASS} Results}
\usage{
\method{print}{CBASS}(x, ...)
}
\arguments{
\item{x}{an object of class \code{CBASS} as returned by \code{\link{CBASS}}}

\item{...}{Additional unused arguments}
}
\description{
Prints a brief descripton of a fitted \code{CBASS} object.
}
\details{
Reports number of observations and variables of dataset, any preprocessing
done by the \code{\link{CBASS}} function, regularization weight information,
the variant of \code{CBASS} used, and the visualizations returned.
}
\examples{
cbass_fit <- CBASS(X=presidential_speech[1:10,1:4])
print(cbass_fit)
}
