% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WeightAdjacency}
\alias{WeightAdjacency}
\title{Construct adjacency matrix induced by weights}
\usage{
WeightAdjacency(weights, nobs, weighted = FALSE, upper = TRUE)
}
\arguments{
\item{weights}{a vector of weights such as returned by \code{SparseWeights}}

\item{nobs}{the number of observations being clustered}

\item{weighted}{a logical. If \code{FALSE} created unweighted adjacency matrix
determined by support of weight vector. If TRUE create weighted adjacency;
default is FALSE.}

\item{upper}{a logical. If \code{TRUE} only return upper triangular of matrix.
If \code{FALSE} return symmetric adjacency}
}
\value{
adj a sparse adjacency matrix
}
\description{
Construct adjacency matrix induced by weights
}
\keyword{internal}
